backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-multipart
PyPDF2
python-docx
spacy
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
pymongo
certifi
python-dotenv
pandas
numpy
openai
bcrypt
email-validator
